Renamed modules to distinguish b/w Bar and StackedBar charts
