+++ src/components/StackedBarChart/StackedBarChart.types.js
