+++ src/components/StackedBarChart/index.js
@@ -1,1 +1,1 @@
-export { default } from './BarChart';
+export { default } from './StackedBarChart';
